refactor(daemon): narrow compatibility turn execution surfaces
